refactor: relocate schema make helpers
