+++ pkgs/standards/autoapi/autoapi/v3/schema/__init__.py
@@ -1,6 +1,13 @@
 # autoapi/v3/schema/__init__.py
 from .builder import _build_schema, _build_list_params
-from .utils import namely_model
+from .utils import (
+    namely_model,
+    _make_bulk_rows_model,
+    _make_bulk_rows_response_model,
+    _make_bulk_ids_model,
+    _make_deleted_response_model,
+    _make_pk_model,
+)
 from .get_schema import get_schema
 from .col_info import (
     VALID_KEYS,
@@ -16,6 +23,11 @@ __all__ = [
     "_build_schema",
     "_build_list_params",
     "namely_model",
+    "_make_bulk_rows_model",
+    "_make_bulk_rows_response_model",
+    "_make_bulk_ids_model",
+    "_make_deleted_response_model",
+    "_make_pk_model",
     "get_schema",
     "VALID_KEYS",
     "VALID_VERBS",
